+++ plugin/src/main/groovy/org/openbakery/XcodeBuildTask.groovy
@@ -49,7 +49,15 @@ class XcodeBuildTask extends AbstractXcodeBuildTask {
 
 		StyledTextOutput output = getServices().get(StyledTextOutputFactory.class).create(XcodeBuildTask.class);
 
-		commandRunner.run("${project.projectDir.absolutePath}", commandList, new XcodeBuildOutputAppender(output))
+		try {
+			commandRunner.run("${project.projectDir.absolutePath}", commandList, new XcodeBuildOutputAppender(output))
+		} finally {
+			String commandOutput = commandRunner.getResult();
+			File outputDirectory = project.getBuildDir();
+			new File(outputDirectory, "xcodebuild-output.txt").withWriter { out ->
+				out.write(commandOutput)
+			}
+		}
 		logger.quiet("Done")
 	}